backend/requirements.txt
fastapi
uvicorn
python-multipart
chromadb
sentence-transformers
groq
python-dotenv
pymupdf